renderAddKeyFromBackup moves to recovery
